Fix logic for determinism for asynchronous tests
More explicit logic for the sequence of the operation has been added.
For this, sleep and sequential executions have been added with the java CompletableFuture Api.
Also, with the help of lua, we will find out exactly which connection will come to us by round-robin, which allows us to avoid implicit behavior.
More sequential logic is done using the callback for the retry api.