Always failing tests have been changed
The tests were always failing, but they showed ok, since the suite ended earlier than the threads in which these tests were run.
Changed the logic of the main application, there was an infinite recursion due to a typo.
Added sleep as threads from other tests affect tests.
Since the tests were run in different threads and used the same resources (spaces), a separate space was made for certain tests.