Exception for not exist field name
We use update/upsert from TarantoolSpace, which updates the space via
iproto. To specify the field to update, use fieldName. Previously, if we
made a mistake in the name, no error occurred and a request was sent to
tarantool where fieldIndex = -1, and this is correct for tarantool
because it uses reverse index, so -1 is the last element in the Tuple.
For now, we'll throw a TarantoolSpaceFieldNotFoundException if the
specified fieldName doesn't exist.

Closes #63

Co-authored-by: Vladimir Rogach <v.rogach@corp.mail.ru>
Co-authored-by: Oleg Kuznetsov <oleg.kuznetsov@corp.mail.ru>