Make TupleOperation immutable
The user creates operations by fieldName, which are subsequently used by
the space.update/upsert functions. After using these functions, a
fieldIndex field appears in operations that was missing. To remedy this
situation, we make TupleOperation non-mutable. And every time we need to
change the fields of these operations, for example, in the
space.update/upsert functions, then we clone these operations.

Co-authored-by: Vladimir Rogach <v.rogach@corp.mail.ru>
Co-authored-by: Oleg Kuznetsov <oleg.kuznetsov@corp.mail.ru>