Use String values and class names in the request signature components
The actual values of the request arguments should not affect the
signature value if they are not String values - this is a special
exception to be able to include server function names and return types
as the signature components. For the other components only the order and
argument types matter for this purpose.