Add batch flushing of requests
Previously the outbound requests were not flushed at all, causing
memory overload as Netty does not flush the requests automatically as
of version 4.1.x. But flushing each request will also affect the
performance significantly, so the built-in batch flushing handler is
added to the connection pipeline together with a config option for
tuning the request batch size.